WIP: windows port scaffolding + macOS shaderc/loader fixes
Not for pushing yet. Linux/macOS build green; windows target compiles the
core libs (down from 497 to ~100 errors) but the compositor + substrate are
incomplete. Checkpoint so parallel osx/windows fix agents branch from one state.

 - windows: OS gate + LLVM backend + zig_dlopen_windows + win32_window +
   vk_win32 + video.zig .win32 mode + build.zig wiring + LLP64/size_t fixes
   + Win32 time/sem shims + warning_va.c varargs shim
 - macOS: shaderc loader tries per-OS names (.dylib); vulkan-1.dll loader
   candidate for windows
 - giz std/debug.zig shmifPanicEmit made windows-safe (separate, in the fork)

+++ src/a12/net/dir_srv_link.zig
@@ -187,7 +187,7 @@ fn requestResource(ns: c_int, res: [*:0]const u8, mode: c_int) c_int {
 
     const status = c.dir_request_resource(
         @ptrCast(&G.shmif_parent_process),
-        @bitCast(@as(c_ulong, @intCast(@as(c_uint, @bitCast(ns))))),
+        @bitCast(@as(usize, @intCast(@as(c_uint, @bitCast(ns))))),
         res,
         mode,
         rep,
